requirements.txt
# requirements.txt

Flask==3.0.3
Flask-Cors==4.0.1
requests==2.32.3
gunicorn==22.0.0
python-dotenv==1.0.1
google-cloud-storage==2.16.0 # NUEVA DEPENDENCIA
werkzeug==3.0.3 # Añadido para secure_filename
